Migrate mustache.js to bash (#2)
Run the expression in a Bash shell instead of using mustache.js. This seems closer to what we actually need in practice to do things like storing the output of a command run in a subshell.

It also allows us to remove all runtime dependencies, thus making the build step redundant.

The action now probably only supports linux and mac. This should be easily fixable if we want to use bash on all platforms (if so the action name should reflect that) by looking up the bash executable based on runner type as it's preinstalled in the windows images as well. But we might also default to powershell on windows, which is probably more reasonable as this is the default shell on windows runners.